Move currency back into type
