+++ server/graphql/general/src/lib.rs
@@ -9,7 +9,10 @@ use chrono::{DateTime, Utc};
 use graphql_core::pagination::PaginationInput;
 
 use crate::store_preference::store_preferences;
-use graphql_types::types::{StorePreferenceNode, TemperatureLogFilterInput};
+use graphql_types::types::{
+    CurrenciesResponse, CurrencyFilterInput, CurrencySortInput, StorePreferenceNode,
+    TemperatureLogFilterInput,
+};
 use mutations::{
     barcode::{insert_barcode, BarcodeInput},
     common::SyncSettingsInput,
